Finishes Donjon Import UI
The UI now accepts both patron image uploads and regular user JSON only uploads. The UI also displays warning messages when certain conditions are hit (i.e. user is out of image slots or not a patron).

Fixes a bug with Encounters master list not resizing as items are opened